feat: CUS-02 기능구현 및 테스트 완료 (#18)
* feat: CUS-02 원두 상세정보 출력 기능 구현 및 테스트 완료

* test: CUS-02 원두 상세 조회 기능 테스트 코드 추가

+++ src/main/java/com/back/orderplz_01/coffee/repository/CoffeeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.back.orderplz_01.coffee.entity.Coffee;
 
-public interface CoffeeRepository extends JpaRepository<Coffee, Integer> {
+public interface CoffeeRepository extends JpaRepository<Coffee, Long> {
 }
